Change return type in bandwidth attribute of network information API https://bugs.webkit.org/show_bug.cgi?id=89959
Reviewed by Adam Barth.

According to network information API spec, bandwidth attribute should return double type.
But, existing implementation has used *long* and *unsigned int* types.

Source/WebCore:

No new tests. Covered by existing tests.

* Modules/networkinfo/NetworkInfo.cpp:
(WebCore::NetworkInfo::NetworkInfo):
* Modules/networkinfo/NetworkInfo.h:
(WebCore::NetworkInfo::create):
(WebCore::NetworkInfo::bandwidth):
(NetworkInfo):
* Modules/networkinfo/NetworkInfoClient.h:
(NetworkInfoClient):
* Modules/networkinfo/NetworkInfoConnection.cpp:
(WebCore::NetworkInfoConnection::bandwidth):
* Modules/networkinfo/NetworkInfoConnection.h:
(NetworkInfoConnection):
* testing/Internals.cpp:
(WebCore::Internals::setNetworkInformation):
* testing/Internals.h:
(Internals):
* testing/Internals.idl:

Source/WebKit/efl:

* WebCoreSupport/NetworkInfoClientEfl.cpp:
(WebCore::NetworkInfoClientEfl::bandwidth):
* WebCoreSupport/NetworkInfoClientEfl.h:
(NetworkInfoClientEfl):


Canonical link: https://commits.webkit.org/107881@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121312 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/networkinfo/NetworkInfo.cpp
@@ -32,7 +32,7 @@
 #if ENABLE(NETWORK_INFO)
 namespace WebCore {
 
-NetworkInfo::NetworkInfo(long bandwidth, bool metered)
+NetworkInfo::NetworkInfo(double bandwidth, bool metered)
     : m_bandwidth(bandwidth)
     , m_metered(metered)
 {
